# resources/openAPI/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Dynamic Mapper API",
    "description" : "REST API for managing Dynamic Mapper configurations including connectors, mappings,\nservice settings, and code templates. The Dynamic Mapper enables bi-directional\ndata transformation between external systems and Cumulocity IoT.\n\n**Key Features:**\n- Connector management (MQTT, HTTP, TCP, etc.)\n- Service configuration and feature flags\n- Code template management for custom processing\n- Multi-tenant support with role-based access control\n",
    "contact" : {
      "name" : "Cumulocity Dynamic Mapper Team",
      "email" : "support@cumulocity.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "v1.0"
  },
...
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "basicAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/extension" : {
      "get" : {
        "tags" : [ "Extension Controller" ],
        "summary" : "Get all processor extensions",
        "description" : "Retrieves all available processor extensions for the current tenant. Extensions provide custom data transformation and processing capabilities that can be used in mappings.",
        "operationId" : "getProcessorExtensions",
        "responses" : {
          "200" : {
            "description" : "Extensions retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Map of extension names to their configurations"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
    "/extension/{extensionName}" : {
      "get" : {
        "tags" : [ "Extension Controller" ],
        "summary" : "Get a specific processor extension",
        "description" : "Retrieves detailed information about a specific processor extension including its configuration, status, and available entry points.",
        "operationId" : "getProcessorExtension",
        "parameters" : [ {
          "name" : "extensionName",
          "in" : "path",
          "description" : "The unique name of the extension to retrieve",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "custom-json-processor"
        } ],
        "responses" : {
          "200" : {
            "description" : "Extension found and retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Extension"
                }
              }
            }
          },
          "404" : {
            "description" : "Extension not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Error details indicating the extension was not found"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ "Extension Controller" ],
        "summary" : "Delete a processor extension",
        "description" : "Deletes a processor extension from the system. This will remove the extension and make it unavailable for use in mappings. Only external extensions can be deleted - built-in extensions cannot be removed.\n\n**Security:** Requires ROLE_DYNAMIC_MAPPER_ADMIN role.\n",
        "operationId" : "deleteProcessorExtension",
        "parameters" : [ {
          "name" : "extensionName",
          "in" : "path",
          "description" : "The unique name of the extension to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "custom-json-processor"
        } ],
        "responses" : {
          "200" : {
            "description" : "Extension deleted successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Extension"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - insufficient permissions to delete extensions"
          },
          "404" : {
            "description" : "Extension not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "description" : "Error details indicating the extension was not found"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ExtensionEntry" : {
        "type" : "object",
        "properties" : {
          "extensionName" : {
            "type" : "string"
          },
          "eventName" : {
            "type" : "string"
          },
          "fqnClassName" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "parameter" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "extensionType" : {
            "type" : "string",
            "enum" : [ "EXTENSION_INBOUND", "EXTENSION_OUTBOUND" ]
          },
          "direction" : {
            "type" : "string",
            "enum" : [ "INBOUND", "OUTBOUND", "UNSPECIFIED" ]
          }
        },
        "required" : [ "direction", "eventName", "extensionName", "extensionType", "fqnClassName" ]
      },
...
      "Extension" : {
        "type" : "object",
        "description" : "Processor extension configuration providing custom data transformation capabilities",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Unique identifier for the extension",
            "example" : "custom-json-processor"
          },
          "loaded" : {
            "type" : "string",
            "description" : "Loading status of the extension",
            "enum" : [ "COMPLETE", "PARTIALLY", "NOT_LOADED", "UNKNOWN" ],
            "example" : "LOADED"
          },
          "name" : {
            "type" : "string",
            "description" : "Display name of the extension",
            "example" : "Custom JSON Processor"
          },
          "external" : {
            "type" : "boolean",
            "description" : "Whether this is an external extension (true) or built-in extension (false)",
            "example" : true
          },
          "extensionEntries" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/ExtensionEntry"
            },
            "description" : "Map of available extension entry points and their configurations"
          }
        },
        "required" : [ "extensionEntries", "external", "id", "loaded", "name" ]
      },
...
    },
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "description" : "Basic Authentication using Cumulocity IoT credentials",
        "in" : "header",
        "scheme" : "basic"
      }
    }
  }
}